+++ psr.c
@@ -10,13 +10,15 @@
 #include "zend_API.h"
 #include "zend_interfaces.h"
 
+#include "php5to7.h"
+
 #include "php_psr.h"
 #include "psr_cache.h"
 #include "psr_container.h"
 #include "psr_http_message.h"
 #include "psr_link.h"
 #include "psr_log.h"
-#include "php5to7.h"
+#include "psr_simple_cache.h"
 
 static PHP_MINIT_FUNCTION(psr)
 {
@@ -25,6 +27,7 @@ static PHP_MINIT_FUNCTION(psr)
     PHP_MINIT(psr_http_message)(INIT_FUNC_ARGS_PASSTHRU);
     PHP_MINIT(psr_link)(INIT_FUNC_ARGS_PASSTHRU);
     PHP_MINIT(psr_log)(INIT_FUNC_ARGS_PASSTHRU);
+    PHP_MINIT(psr_simple_cache)(INIT_FUNC_ARGS_PASSTHRU);
     return SUCCESS;
 }
 
@@ -39,6 +42,7 @@ static PHP_MINFO_FUNCTION(psr)
     php_info_print_table_row(2, "PSR-7 Http Message Version", PHP_PSR_HTTP_MESSAGE_VERSION);
     php_info_print_table_row(2, "PSR-11 Container Version", PHP_PSR_CONTAINER_VERSION);
     php_info_print_table_row(2, "PSR-13 Link Version", PHP_PSR_LOG_VERSION);
+    php_info_print_table_row(2, "PSR-16 Simple Cache Version", PHP_PSR_SIMPLE_CACHE_VERSION);
     php_info_print_table_end();
 }